feat: reorganize the package and fix issues
